# Conflicts:
#	.github/workflows/docker_pr_test.yml
#	backend/lib/manager.py
#	backend/lib/processor.py
#	backend/workers/restart_4cat.py
#	common/lib/dataset.py
#	datasources/bitchute/search_bitchute.py
#	datasources/bsky/search_bsky.py
#	datasources/fourchan/scrapers/scrape_threads.py
#	datasources/fourchan/search_4chan.py
#	datasources/reddit/search_reddit.py
#	datasources/telegram/search_telegram.py
#	datasources/tumblr/search_tumblr.py
#	processors/audio/audio_extractor.py
#	processors/filtering/write_annotations.py
#	processors/machine_learning/openai_prompt.py
#	processors/machine_learning/text_from_image.py
#	processors/metrics/overtime-hatebase.py
#	processors/metrics/rank_attribute.py
#	processors/presets/monthly-histogram.py
#	processors/text-analysis/tokenise.py
#	processors/visualisation/download_images.py
#	processors/visualisation/image_category_wall.py
#	processors/visualisation/image_wall_w_text.py
#	processors/visualisation/video_hasher.py
#	processors/visualisation/video_stack.py
#	webtool/__init__.py
#	webtool/lib/template_filters.py
#	webtool/templates/components/result-details.html
#	webtool/views/api_explorer.py
#	webtool/views/api_tool.py
#	webtool/views/views_admin.py
#	webtool/views/views_dataset.py
#	webtool/views/views_extensions.py
#	webtool/views/views_restart.py
#	webtool/views/views_user.py

+++ backend/workers/cleanup_tempfiles.py
@@ -98,9 +98,12 @@ class TempFileCleaner(BasicWorker):
                 # if the dataset is finished, the staging area should have been
                 # compressed into a zip file, or deleted, so this is also safe
                 # to clean up
-                self.log.info("Dataset %s is finished, but staging area remains at %s, deleting folder" % (
-                dataset.key, str(file)))
-                shutil.rmtree(file)
+                if file.name not in tracked_files:
+                    self.log.info("Dataset %s is finished, but staging area remains at %s, marking for deletion" % (dataset.key, str(file)))
+                    tracked_files[file.name] = datetime.now().timestamp() + (self.days_to_keep * 86400)
+                elif tracked_files[file.name] < datetime.now().timestamp():
+                    self.log.info("Dataset %s is finished, but staging area remains at %s, deleting folder" % (dataset.key, str(file)))
+                    shutil.rmtree(file)
 
         # Update tracked files
         tracking_file.write_text(json.dumps(tracked_files))
